auth routes

+++ server/src/routes/memories.ts
@@ -3,8 +3,14 @@ import { z } from 'zod'
 import { prisma } from '../lib/prisma'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
+  app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: { createdAt: 'asc' },
     })
 
@@ -17,7 +23,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     const parmsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -30,6 +36,9 @@ export async function memoriesRoutes(app: FastifyInstance) {
       },
     })
 
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
     return memory
   })
 
@@ -42,17 +51,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.create({
+    await prisma.memory.create({
       data: {
         content,
         coverUrl,
         isPublic,
-        userId: '71ff9c16-e9c6-4814-a9f0-acf5e3540fdd',
+        userId: request.user.sub,
       },
     })
   })
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const parmsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -67,7 +76,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findFirstOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: { id },
       data: { content, coverUrl, isPublic },
     })
@@ -75,14 +94,23 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory
   })
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const parmsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = parmsSchema.parse(request.params)
 
-    const memory = await prisma.memory.delete({
+    const memory = await prisma.memory.findFirstOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+    await prisma.memory.delete({
       where: {
         id,
       },
